move jwt functions in their own service
