--- frontend/src/services/apiService.ts
@@ -1,1 +1,0 @@
-export const x =0;
