+++ frontend/src/services/employeeService.ts
@@ -1,0 +1,3 @@
+import axios from "axios";
+import {ICredentials} from "../interfaces/IEmployee";
+import {parseError} from './errorService';
